add attach data to the node models.
